backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
boto3==1.34.34
chromadb==0.5.23
pydantic==2.5.3
python-dotenv==1.0.0
